trading_agent/requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0

# Trading Platform APIs
alpaca-py>=0.8.0  # Alpaca trading API
yfinance>=0.2.0   # Yahoo Finance for backup data

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For parquet file support

# Messaging
slack-bolt>=1.18.0
slack-sdk>=3.23.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations

# Async Support
aiofiles>=23.0.0
aiohttp>=3.9.0

# Technical Analysis
ta-lib>=0.4.28  # Technical indicators (optional, needs system install)
pandas-ta>=0.3.14b  # Pure Python alternative

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0